pyproject.toml
[tool.poetry]
name = "valar-morghulis"
version = "0.1.0"
description = ""
authors = ["Pierre-Louis Sergent <papa.louis59@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.11.3"
fastapi = "^0.54.1"
requests = "^2.24.0"
passlib = "^1.7.2"
fastapi_sqlalchemy = "^0.1.4"
email_validator = "^1.1.1"
psycopg2 = "^2.8.5"
bcrypt = "^3.1.7"
python-multipart = "^0.0.5"
alembic = "^1.4.2"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"
pdbpp = "^0.10.2"
alembic = "^1.4.2"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

